- Changing WU count in DB from long to int - Fixed issue where not retrieving all users from DB if cache is evicted

+++ folding-stats-api/src/main/java/me/zodac/folding/api/UserStats.java
@@ -8,9 +8,9 @@ import java.util.Objects;
 public class UserStats {
 
     private final long points;
-    private final long wus;
+    private final int wus;
 
-    public UserStats(final long points, final long wus) {
+    public UserStats(final long points, final int wus) {
         this.points = points;
         this.wus = wus;
     }
@@ -19,7 +19,7 @@ public class UserStats {
         return points;
     }
 
-    public long getWus() {
+    public int getWus() {
         return wus;
     }
